+++ Sources/Sentry/SentryScreenshotIntegration.m
@@ -1,6 +1,7 @@
 #import "SentryScreenshotIntegration.h"
 #import "SentryAttachment.h"
 #import "SentryClient+Private.h"
+#import "SentryCrashC.h"
 #import "SentryDependencyContainer.h"
 #import "SentryEvent+Private.h"
 #import "SentryEvent.h"
@@ -10,6 +11,35 @@
 #import "SentrySDK+Private.h"
 
 #if SENTRY_HAS_UIKIT
+
+void
+saveScreenShot(const char *path)
+{
+    NSString *reportPath = [NSString stringWithUTF8String:path];
+    NSError *error = nil;
+
+    if (![NSFileManager.defaultManager fileExistsAtPath:reportPath]) {
+        [NSFileManager.defaultManager createDirectoryAtPath:reportPath
+                                withIntermediateDirectories:YES
+                                                 attributes:nil
+                                                      error:&error];
+        if (error != nil)
+            return;
+    } else {
+        // We first delete any screenshot that could be from an old crash report
+        NSArray *oldFiles = [NSFileManager.defaultManager contentsOfDirectoryAtPath:reportPath
+                                                                              error:&error];
+
+        if (!error) {
+            [oldFiles enumerateObjectsUsingBlock:^(NSString *obj, NSUInteger idx, BOOL *stop) {
+                [NSFileManager.defaultManager removeItemAtPath:obj error:nil];
+            }];
+        }
+    }
+
+    [SentryDependencyContainer.sharedInstance.screenshot saveScreenShots:reportPath];
+}
+
 @implementation SentryScreenshotIntegration
 
 - (void)installWithOptions:(nonnull SentryOptions *)options
@@ -21,6 +51,13 @@
 
     SentryClient *client = [SentrySDK.currentHub getClient];
     client.attachmentProcessor = self;
+
+    sentrycrash_setSaveScreenshots(&saveScreenShot);
+}
+
+- (void)uninstall
+{
+    sentrycrash_setSaveScreenshots(NULL);
 }
 
 - (NSArray<SentryAttachment *> *)processAttachments:(NSArray<SentryAttachment *> *)attachments
@@ -39,11 +76,8 @@
     [result addObjectsFromArray:attachments];
 
     for (int i = 0; i < screenshot.count; i++) {
-        NSString *name;
-        if (i == 0)
-            name = @"screenshot.png";
-        else
-            name = [NSString stringWithFormat:@"screenshot-%i.png", i + 1];
+        NSString *name
+            = i == 0 ? @"screenshot.png" : [NSString stringWithFormat:@"screenshot-%i.png", i + 1];
 
         SentryAttachment *att = [[SentryAttachment alloc] initWithData:screenshot[i]
                                                               filename:name