+++ Sources/SentryCrash/Recording/SentryCrashC.h
@@ -133,6 +133,13 @@ void sentrycrash_setPrintPreviousLog(bool shouldPrintPreviousLog);
  */
 void sentrycrash_setMaxReportCount(int maxReportCount);
 
+/**
+ * Set the callback to be called at the end of a crash to make the app save a screenshot;
+ *
+ * @param callback function pointer that will be called with a give path.
+ */
+void sentrycrash_setSaveScreenshots(void (*callback)(const char *));
+
 /** Report a custom, user defined exception.
  * This can be useful when dealing with scripting languages.
  *
@@ -227,6 +234,12 @@ void sentrycrash_deleteAllReports(void);
  * @param reportID An ID of report to delete.
  */
 void sentrycrash_deleteReportWithID(int64_t reportID);
+
+/**
+ * For testing purpose.
+ * Indicates that a callback was registered for screenshot.
+ */
+bool sentrycrash_hasSaveScreenshotCallback(void);
 
 #ifdef __cplusplus
 }